Solutions

GeometricIntegrators.Solutions.AtomicSolution — Type

Abstract atomistic or single-step solution.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for DAE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for ODE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for partitioned DAE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for partitioned ODE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Print error for AtomicSolutions of equations not implemented, yet.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for DAE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for ODE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for partitioned DAE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Create AtomicSolution for partitioned ODE.

GeometricIntegrators.Solutions.AtomicSolution — Method

Print error for AtomicSolutions of solution not implemented, yet.

GeometricIntegrators.Solutions.AtomicSolutionDAE — Type

Atomic solution for an DAE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̅: time of previous time step

q: current solution of q

q̅: previous solution of q

q̃: compensated summation error of q

λ: current solution of λ

λ̅: previous solution of λ

v: vector field of q

v̅: vector field of q̅

u: projective vector field of q

u̅: projective vector field of q̅

GeometricIntegrators.Solutions.AtomicSolutionODE — Type

Atomic solution for an ODE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̅: time of previous time step

q: current solution of q

q̅: previous solution of q

q̃: compensated summation error of q

v: vector field of q

v̅: vector field of q̅

GeometricIntegrators.Solutions.AtomicSolutionPDAE — Type

Atomic solution for an PDAE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̅: time of previous time step

q: current solution of q

q̅: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̅: previous solution of p

p̃: compensated summation error of p

λ: current solution of λ

λ̅: previous solution of λ

v: vector field of q

v̅: vector field of q̅

f: vector field of p

f̅: vector field of p̅

u: projective vector field of q

u̅: projective vector field of q̅

g: projective vector field of p

g̅: projective vector field of p̅

GeometricIntegrators.Solutions.AtomicSolutionPODE — Type

Atomic solution for an PODE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̅: time of previous time step

q: current solution of q

q̅: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̅: previous solution of p

p̃: compensated summation error of p

v: vector field of q

v̅: vector field of q̅

f: vector field of p

f̅: vector field of p̅

GeometricIntegrators.Solutions.Solution — Method

Create solution for DAE.

GeometricIntegrators.Solutions.Solution — Method

Create solution for ODE.

GeometricIntegrators.Solutions.Solution — Method

Create solution for partitioned DAE.

GeometricIntegrators.Solutions.Solution — Method

Create solution for partitioned ODE.

GeometricIntegrators.Solutions.Solution — Method

Print error for solutions of equations not implemented, yet.

GeometricIntegrators.Solutions.SolutionDAE — Type

SolutionDAE: Solution of a differential algebraic equation

Contains all fields necessary to store the solution of an DAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

GeometricIntegrators.Solutions.SolutionODE — Type

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

GeometricIntegrators.Solutions.SolutionPDAE — Type

SolutionPDAE: Solution of a partitioned differential algebraic equation

Contains all fields necessary to store the solution of an PDAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

GeometricIntegrators.Solutions.SolutionPODE — Type

SolutionPODE: Solution of a partitioned ordinary differential equation

Contains all fields necessary to store the solution of an PODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

GeometricIntegrators.Common.write_to_hdf5 — Function

Append solution to HDF5 file.

GeometricIntegrators.Common.write_to_hdf5 — Function

Append solution to HDF5 file.

GeometricIntegrators.Common.write_to_hdf5 — Method

Creates HDF5 file, writes solution to file, and closes file.

GeometricIntegrators.Common.write_to_hdf5 — Method

writetohdf5: Wrapper for saving Solution to HDF5 file.

GeometricIntegrators.Solutions.ParallelSolution — Method

Create parallel solution for DAE.

GeometricIntegrators.Solutions.ParallelSolution — Method

Create parallel solution for ODE.

GeometricIntegrators.Solutions.ParallelSolution — Method

Create parallel solution for partitioned DAE.

GeometricIntegrators.Solutions.ParallelSolution — Method

Create parallel solution for partitioned ODE.

GeometricIntegrators.Solutions.ParallelSolution — Method

Print error for parallel solutions of equations not implemented, yet.

GeometricIntegrators.Solutions.copy_solution! — Method

Copy solution from atomistic solution to solution object.

GeometricIntegrators.Solutions.createHDF5 — Method

createHDF5: Creates or opens HDF5 file.

GeometricIntegrators.Solutions.create_hdf5! — Method

Creates HDF5 file and initialises datasets for solution object.

GeometricIntegrators.Solutions.create_hdf5 — Method

Creates HDF5 file and initialises datasets for deterministic solution object.

GeometricIntegrators.Solutions.save_attributes — Method

save_attributes: Saves attributes of Deterministic Solutions to HDF5 file.

GeometricIntegrators.Solutions.save_attributes — Method

save_attributes: Saves common attributes of Solution to HDF5 file.