Runge-Kutta Tableaus

RungeKutta.PartitionedTableau — Type

Tableau of a Partitioned Runge-Kutta method

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, v(t_{n} + c_j \Delta t, Q_{n,j}, P_{n,j}) , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, v(t_{n} + c_j \Delta t, Q_{n,i}, P_{n,i}) , \\ P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, f(t_{n} + c_j \Delta t, Q_{n,j}, P_{n,j}) , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, f(t_{n} + c_j \Delta t, Q_{n,i}, P_{n,i}) . \end{aligned}\]

Parameters:

T: datatype of coefficient arrays

Fields:

name: symbolic name of the tableau

o: order of the method

s: number of stages

q: Tableau for q

p: Tableau for p

The actual tableaus are stored in q and p:

a: coefficients $a_{ij}$ with $ 1 \le i,j \le s$

b: weights $b_{i}$ with $ 1 \le i \le s$

c: nodes $c_{i}$ with $ 1 \le i \le s$

Constructors:

PartitionedTableau{T}(name, o, s, q, p)
PartitionedTableau{T}(name, q, p)
PartitionedTableau(name::Symbol, q::Tableau, p::Tableau)
PartitionedTableau(name::Symbol, q::Tableau)

RungeKutta.Tableau — Type

Holds the tableau of a Runge-Kutta method

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, v(t_{n} + c_j \Delta t, Q_{n,j}) , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, v(t_{n} + c_j \Delta t, Q_{n,i}) , \\ \end{aligned}\]

Parameters:

T: datatype of coefficient arrays

Fields:

name: symbolic name of the tableau

o: order of the method

s: number of stages

a: coefficients $a_{ij}$ with $ 1 \le i,j \le s$

b: weights $b_{i}$ with $ 1 \le i \le s$

c: nodes $c_{i}$ with $ 1 \le i \le s$

Constructors:

Tableau{T}(name, o, s, a, b, c)
Tableau{T}(name, o, a, b, c)
Tableau(name::Symbol, o::Int, s::Int, a::AbstractMatrix, b::AbstractVector, c::AbstractVector)
Tableau(name::Symbol, o::Int, a::AbstractMatrix, b::AbstractVector, c::AbstractVector)
Tableau(name::Symbol, o::Int, t::AbstractMatrix)

The last constructor accepts an $(s+1) \times (s+1)$ array that holds the whole tableau in the form of a Butcher tableau, i.e.,

c	a

b

Base.show — Method

Base.show(io::IO, ::MIME"text/markdown", tab::Tableau)

Generate and print a nice markdown table for the Runge-Kutta tableau.

Base.show — Method

Base.show(io::IO, tab::Tableau)

Pretty-print Runge-Kutta tableau.

RungeKutta._legendre — Method

Legendre polynomial P_s(x) of degree s defined on the interval [-1..+1].

RungeKutta._shifted_legendre — Function

Legendre polynomial of degree s shifted to the interval [0..1], i.e., P_s(2x-1).

RungeKutta.from_file — Method

from_file(dir::AbstractString, name::AbstractString)

Read Runge-Kutta tableau from the file <name>.tsv in the directory dir.

RungeKutta.solve_simplifying_assumption_b — Method

Compute the weights by solving the simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_c — Method

Compute the coefficients by solving the simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_d — Method

Compute the coefficients by solving the simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.to_file — Method

to_file(dir::AbstractString, tab::Tableau)

Write Runge-Kutta tableau to the file <tab.name>.tsv in the directory dir.