Solutions

Solutions

GeometricIntegrators.Solutions.Solution — Type.

Create solution for variational ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for SDE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for PSDE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for ODE and split ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Print error for solutions of equations not implemented, yet.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit DAE.

source

GeometricIntegrators.Solutions.SolutionDAE — Type.

SolutionDAE: Solution of a differential algebraic equation

Contains all fields necessary to store the solution of an DAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

source

GeometricIntegrators.Solutions.SolutionODE — Type.

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

source

GeometricIntegrators.Solutions.SolutionPDAE — Type.

SolutionPDAE: Solution of a partitioned differential algebraic equation

Contains all fields necessary to store the solution of an PDAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

source

GeometricIntegrators.Solutions.SolutionPODE — Type.

SolutionPODE: Solution of a partitioned ordinary differential equation

Contains all fields necessary to store the solution of an PODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

source

GeometricIntegrators.Solutions.SolutionPSDE — Type.

SolutionPSDE: Solution of a partitioned stochastic differential equation

Contains all fields necessary to store the solution of a PSDE or SPSDE

Fields

conv: type of the solution: "strong" or "weak"

nd: dimension of the dynamical variable $q$

nm: dimension of the Wiener process

nt: number of time steps to store

ns: number of sample paths

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ns, ni] with q[:,0,:,:] the initial conditions

p: solution p[nd, nt+1, ns, ni] with p[:,0,:,:] the initial conditions

W: Wiener process driving the stochastic processes q and p

K: integer parameter defining the truncation of the increments of the Wiener process (for strong solutions),

  A = √(2 K Δt |log Δt|) due to Milstein & Tretyakov; if K=0 no truncation

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Solutions.SolutionSDE — Type.

SolutionSDE: Solution of a stochastic differential equation

Contains all fields necessary to store the solution of an SDE.

Fields

conv: type of the solution: "strong" or "weak"

nd: dimension of the dynamical variable $q$

nm: dimension of the Wiener process

nt: number of time steps to store

ns: number of sample paths

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ns, ni] with q[:,0,:,:] the initial conditions

W: Wiener process driving the stochastic process q

K: integer parameter defining the truncation of the increments of the Wiener process (for strong solutions),

  A = √(2 K Δt |log Δt|) due to Milstein & Tretyakov; if K=0 no truncation

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for SDE solution object. It is implemented as one fucntion for all NQ and NW cases, rather than several separate cases as was done for SolutionODE. nt - the total number of time steps to store ntime - the total number of timesteps to be computed

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for SDE solution object. It is implemented as one function for all NQ and NW cases, rather than several separate cases as was done for SolutionODE. nt - the total number of time steps to store ntime - the total number of timesteps to be computed

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with single initial condition.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with multiple initial conditions.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.get_initial_conditions! — Method.

copies the m-th initial condition for the k-th sample path from sol.q to q

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Creates HDF5 file, writes solution to file, and closes file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

writetohdf5: Wrapper for saving Solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file. offset - start writing q at the position offset+2 offset2- start writing ΔW, ΔZ at the position offset2+1

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file. offset - start writing q at the position offset+2 offset2- start writing ΔW, ΔZ at the position offset2+1

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.Solutions.createHDF5 — Method.

createHDF5: Creates or opens HDF5 file.

source

GeometricIntegrators.Solutions.save_attributes — Method.

save_attributes: Saves attributes of Deterministic Solutions to HDF5 file.

source

GeometricIntegrators.Solutions.save_attributes — Method.

save_attributes: Saves common attributes of Solution to HDF5 file.

source

GeometricIntegrators.Solutions.save_attributes — Method.

save_attributes: Saves attributes of Stochastic Solutions to HDF5 file.

source